fix di for console prompter
