+++ ynac.cli/BudgetSelection/IConsolePrompt.cs
@@ -1,0 +1,8 @@
+using Spectre.Console;
+
+namespace ynac.BudgetSelection;
+
+public interface IConsolePrompt
+{
+    T Prompt<T>(SelectionPrompt<T> prompt) where T : notnull;
+}
