+++ ynac.cli/BudgetSelection/IPrompter.cs
@@ -1,0 +1,6 @@
+namespace ynac.BudgetSelection;
+
+public interface IPrompter
+{
+    T PromptSelection<T>(IReadOnlyCollection<T> items, string title, Func<T, string> converter) where T : notnull;
+}
